Updated Solution
### Since Last Commit
---
---
#### Additions
- N/A
---
#### Changes
- Spawn coordinated from Floats to Integers in Playable Character and Enemy classes.
---
#### Removed
- N/A

+++ vsSolutions/shadowOperative_001/shadowOperative_001/Enemy.cpp
@@ -8,7 +8,7 @@
 
 #include "Enemy.h"
 #include "TextureHolder.h"
-void Enemy::spawn(sf::Vector2f startPosition, float gravity)
+void Enemy::spawn(sf::Vector2i startPosition, float gravity)
 {
 	m_SpawnPosition = startPosition;
 	m_Position = m_SpawnPosition;
